requirements.txt
openai
openpyxl
requests
pandas
beautifulsoup4
numpy
uvicorn
fastapi
rich
python-dotenv
fastapi-cors
chardet